% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build.similarity.matrix}
\alias{build.similarity.matrix}
\title{Build Similarity Matrix}
\usage{
build.similarity.matrix(labels, number.labels)
}
\arguments{
\item{labels}{A data frame representing the label space.}

\item{number.labels}{An integer specifying the number of labels.}
}
\value{
A matrix initialized with zeros and with proper row and column names.
}
\description{
This function initializes an empty similarity matrix for a given number of labels.
}
\examples{
labels <- data.frame(matrix(sample(0:1, 100, replace = TRUE), ncol = 10))
build.similarity.matrix(labels, ncol(labels))
}
