% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_up}
\alias{clean_up}
\title{Clean Up Temporary Files}
\usage{
clean_up(parameters)
}
\arguments{
\item{directories}{A list containing folder paths.}
}
\value{
None. Deletes specified directories.
}
\description{
Deletes temporary result folders and triggers garbage collection.
}
