% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.ab}
\alias{compute.ab}
\title{Compute AB: Sum of A and B}
\usage{
\method{compute}{ab}(a, b)
}
\arguments{
\item{a}{Numeric value representing shared 1s.}

\item{b}{Numeric value representing x absent, y present.}
}
\value{
The sum of A and B.
}
\description{
Computes the sum of A (shared 1s) and B (x absent, y present).
}
