% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.ac}
\alias{compute.ac}
\title{Compute AC: Sum of A and C}
\usage{
\method{compute}{ac}(a, c)
}
\arguments{
\item{a}{Numeric value representing shared 1s.}

\item{c}{Numeric value representing y absent, x present.}
}
\value{
The sum of A and C.
}
\description{
Computes the sum of A (shared 1s) and C (y absent, x present).
}
