% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.ad}
\alias{compute.ad}
\title{Compute AD: Sum of A and D}
\usage{
\method{compute}{ad}(a, d)
}
\arguments{
\item{a}{Numeric value representing shared 1s.}

\item{d}{Numeric value representing shared 0s.}
}
\value{
The sum of A and D.
}
\description{
Computes the sum of A (shared 1s) and D (shared 0s).
}
