% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.bc}
\alias{compute.bc}
\title{Compute BC: Sum of B and C}
\usage{
\method{compute}{bc}(b, c)
}
\arguments{
\item{b}{Numeric value representing x absent, y present.}

\item{c}{Numeric value representing y absent, x present.}
}
\value{
The sum of B and C.
}
\description{
Computes the sum of B (x absent, y present) and C (y absent, x present).
}
