% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.bd}
\alias{compute.bd}
\title{Compute BD: Sum of B and D}
\usage{
\method{compute}{bd}(b, d)
}
\arguments{
\item{b}{Numeric value representing x absent, y present.}

\item{d}{Numeric value representing shared 0s.}
}
\value{
The sum of B and D.
}
\description{
Computes the sum of B (x absent, y present) and D (shared 0s).
}
