% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.cd}
\alias{compute.cd}
\title{Compute CD: Sum of C and D}
\usage{
\method{compute}{cd}(c, d)
}
\arguments{
\item{c}{Numeric value representing y absent, x present.}

\item{d}{Numeric value representing shared 0s.}
}
\value{
The sum of C and D.
}
\description{
Computes the sum of C (y absent, x present) and D (shared 0s).
}
