% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.contingency.table}
\alias{compute.contingency.table}
\title{Compute Contingency Table (Optimized)}
\usage{
\method{compute}{contingency.table}(labels, number.labels, FUN)
}
\arguments{
\item{labels}{A data frame containing the label space.}

\item{number.labels}{An integer representing the number of labels.}

\item{FUN}{A function that computes a similarity or contingency measure between two vectors.}
}
\value{
A list containing the contingency table as a matrix.
}
\description{
This function calculates a contingency table based on a given function.
It uses parallel computing to improve efficiency.
}
\examples{
labels <- data.frame(matrix(sample(0:1, 100, replace = TRUE), ncol = 10))
compute.contingency.table(labels, ncol(labels), function(x, y) sum(x == y))
}
