% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.mn}
\alias{compute.mn}
\title{Compute MN Matrix}
\usage{
\method{compute}{mn}(labels, number.labels, ma, mb, mc, md)
}
\arguments{
\item{labels}{A data frame representing the label space.}

\item{number.labels}{An integer representing the number of labels.}

\item{ma}{A numeric matrix representing computed values of \code{compute.a}.}

\item{mb}{A numeric matrix representing computed values of \code{compute.b}.}

\item{mc}{A numeric matrix representing computed values of \code{compute.c}.}

\item{md}{A numeric matrix representing computed values of \code{compute.d}.}
}
\value{
A list containing the computed MN matrix.
}
\description{
Computes the MN matrix using four input matrices (\code{ma}, \code{mb}, \code{mc}, \code{md}).
}
\examples{
labels <- matrix(sample(0:1, 100, replace = TRUE), ncol = 10)
ma <- matrix(runif(100), ncol = 10)
mb <- matrix(runif(100), ncol = 10)
mc <- matrix(runif(100), ncol = 10)
md <- matrix(runif(100), ncol = 10)
compute.mn(labels, ncol(labels), ma, mb, mc, md)
}
