% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.n}
\alias{compute.n}
\title{Compute N: Total Elements}
\usage{
\method{compute}{n}(a, b, c, d)
}
\arguments{
\item{a}{Numeric value representing shared 1s.}

\item{b}{Numeric value representing x absent, y present.}

\item{c}{Numeric value representing y absent, x present.}

\item{d}{Numeric value representing shared 0s.}
}
\value{
The total sum of all elements (A, B, C, and D).
}
\description{
Computes the total number of elements in the dataset.
}
