% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folder}
\alias{create_folder}
\title{Create Folder if Not Exists}
\usage{
create_folder(folderResults)
}
\arguments{
\item{folderResults}{A character string specifying the folder path.}
}
\value{
None. The folder is created if it does not exist.
}
\description{
This function creates a folder if it does not already exist.
}
