% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{directories}
\alias{directories}
\title{Create Necessary Directories for the Project}
\usage{
directories(parameters)
}
\arguments{
\item{parameters}{A list containing the key \code{Folder.Results}, which is the root directory for saving results.}
}
\value{
A list with paths to all created directories.
}
\description{
This function creates and manages all required directories for the project, ensuring that
the necessary folder structure exists before running further processes.
}
\examples{
params <- list(Folder.Results = "/tmp/birds")
directories(params)

}
