% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{final_cleanup}
\alias{final_cleanup}
\title{Final Cleanup}
\usage{
final_cleanup(directories)
}
\arguments{
\item{directories}{A list containing folder paths to be deleted.}
}
\value{
None. Deletes specified directories.
}
\description{
Removes result folders and runs garbage collection.
}
