% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_directories}
\alias{get_directories}
\title{Get Directories}
\usage{
get_directories(parameters)
}
\arguments{
\item{parameters}{A list containing configuration settings.}
}
\value{
A list of directories.
}
\description{
Retrieves and returns the directory structure based on given parameters.
}
