% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infoDataSet}
\alias{infoDataSet}
\title{Info Data Set Function}
\usage{
infoDataSet(dataset)
}
\arguments{
\item{dataset}{A dataset object containing various attributes. The dataset should have the following columns:
\itemize{
\item \code{ID}: Identifier for the dataset.
\item \code{Name}: Name of the dataset.
\item \code{Domain}: Domain of the dataset.
\item \code{Instances}: Number of instances in the dataset.
\item \code{Attributes}: Total number of attributes in the dataset (input + output).
\item \code{Labels}: Number of labels in the dataset.
\item \code{Inputs}: Number of input attributes in the dataset.
\item \code{Cardinality}: The label cardinality of the dataset.
\item \code{Density}: Label Density of the dataset.
\item \code{LabelsSets}: The sets of labels associated with the dataset.
\item \code{Single}: Boolean indicating if the dataset is single.
\item \code{MaxFreq}: Maximum frequency of the dataset.
\item \code{MeanIR}: The average imbalance ratio (MeanIR) of the dataset, which measures how imbalanced the dataset is.
\item \code{Scumble}: Some specific metric related to the dataset.
\item \code{TCS}: The TCS value of the dataset.
\item \code{AttStart}: Starting index for the attributes.
\item \code{AttEnd}: Ending index for the attributes.
\item \code{LabelStart}: Starting index for the labels.
\item \code{LabelEnd}: Ending index for the labels.
\item \code{Distinct}: Distinct labels in the dataset.
\item \code{xn}: xdim for Kohonen.
\item \code{yn}: ydim for Kohonen.
\item \code{gridn}: Grid value for Kohonen.
\item \code{max.neighbor}: The maximum number of neighbors for the dataset.
\item \code{LabelDependency}: Dependency between labels in the dataset.
}}
}
\value{
A list containing all the relevant information from the dataset, including IDs, names, attributes, and various dataset-specific metrics.
}
\description{
This function retrieves specific information from a dataset, typically stored in a "datasets-hpmlk.csv" file.
}
\examples{
# Example usage:
dataset <- list(ID = "1", Name = "Dataset1", Domain = "text",
                Instances = 100, Attributes = "23", Labels = 3,
                Inputs = 20, Cardinality = 5, Density = 0.9,
                LabelsSets = 5, Single = TRUE, MaxFreq = 10,
                MeanIR = 0.75, Scumble = 1.2, TCS = 2.3,
                AttStart = 1, AttEnd = 20, LabelStart = 21,
                LabelEnd = 23, Distinct = 10,
                xn = 4, yn = 4, gridn = 16,
                max.neighbor = 22, LabelDependency = 0.89)
result <- infoDataSet(dataset)
result$ID
result$Name
result$max.neighbor
result$LabelDependency

}
