% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labelSpace}
\alias{labelSpace}
\title{Label Space Function}
\usage{
labelSpace(parameters)
}
\arguments{
\item{parameters}{A list containing:
\itemize{
\item \code{Folders$folderCVTR}: Folder path containing the training data.
\item \code{ds$LabelStart, ds$LabelEnd}: Column indices specifying the range of labels.
\item \code{dataset_name}: Name of the dataset (used to create file names).
\item \code{number_folds}: Number of folds for cross-validation.
\item \code{folderResults}: Folder path where results will be saved.
}}
}
\value{
A list with two elements:
\itemize{
\item \code{NamesLabels}: Names of the label columns.
\item \code{Classes}: A list containing the label spaces for each fold.
}
}
\description{
This function separates the label space from the rest of the data to be used for calculating correlations.
}
\examples{
# Example usage:
parameters <- list(Folders = list(folderCVTR = "path/to/folder"), ds = list(LabelStart = 1, LabelEnd = 2), 
                   dataset_name = "dataset", number_folds = 5, folderResults = "path/to/results")
result <- labelSpace(parameters)
result$NamesLabels
result$Classes

}
