% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dataset}
\alias{process_dataset}
\title{Process Dataset}
\usage{
process_dataset(datasets, parameters)
}
\arguments{
\item{datasets}{A dataframe containing dataset information.}

\item{parameters}{A list containing configuration settings, including paths.}
}
\value{
None. Stops execution if the dataset file is missing.
}
\description{
Checks if the dataset file exists, and if so, proceeds with copying and extracting it.
}
