% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_locally}
\alias{save_locally}
\title{Save Locally}
\usage{
save_locally(parameters)
}
\arguments{
\item{parameters}{A list containing paths and similarity settings.}
}
\value{
None. Saves the results locally and compresses them into a tar.gz file.
}
\description{
Saves similarity results to a local folder and compresses them.
}
