% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_runtime}
\alias{save_runtime}
\title{Save Runtime Information}
\usage{
save_runtime(timeTCP, parameters)
}
\arguments{
\item{timeTCP}{A system.time() object containing execution time.}

\item{parameters}{A list containing the folder path to save the file.}
}
\value{
None. The runtime is saved to "Runtime.csv".
}
\description{
Saves the execution time of a process to a CSV file.
}
