% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_to_storage}
\alias{save_to_storage}
\title{Save Results to Storage}
\usage{
save_to_storage(parameters)
}
\arguments{
\item{parameters}{A list containing storage configuration settings.}
}
\value{
None. Executes appropriate storage-saving logic.
}
\description{
Saves results either to cloud storage or locally based on configuration.
}
