% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_options}
\alias{set_options}
\title{Set Global Options}
\usage{
set_options()
}
\value{
None. Options are set globally in R.
}
\description{
Configures global options, including Java memory allocation, error message display,
and scientific notation handling.
}
